refactor(plugins): remove lazy imports + the one real import cycle
Lazy function-local imports (with `# noqa: PLC0415`) were used to dodge
circular imports between the plugin framework and its consumers. Most
weren't real cycles — `app.plugins` is already a leaf (imports no app.*
at top) and `builtin_plugin` is loaded dynamically (the registry pattern
that breaks the only framework↔plugin edge), so those imports can sit at
the top. Fixed from design instead of suppressing:

- Extract `parse_skill_gates` / `skill_name_from_read` / `find_skill_md`
  into a new leaf module `app/ai/skill_gate_utils.py` (stdlib + app.config
  only). This breaks the ONE real cycle: `stop_gates` needed those
  helpers from the heavy `claude_backend_utils` (which imports bash_safety,
  the DB, models, the workspace manager). Now both import the leaf.
- Lift every other lazy `from app.plugins import registered_*` to the
  top (skills_sync, bash_safety, task_runner, browser/manager,
  stop_gates) and move builtin_plugin's gate/guard/backend imports to the
  module top. Drop the redundant lazy `import importlib` in plugins.py.
- Repoint claude_backend_hooks + test_skill_gates at the leaf.

13 of 14 `# noqa: PLC0415` removed. The one kept (`cli.py` uvicorn) is a
deliberate deferred-heavy-import — `vibe-seller stop`/`upgrade` must not
drag uvicorn + FastAPI + the whole app graph onto the path — not a cycle.

`import app.main` succeeds (no circular import); ruff clean; 1079 unit
tests pass.
